# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Parenting Platform Auth API
  description: 보호자·파트너·어드민이 공존하는 육아 플랫폼 (B2C) 의 통합 인증/인가 설계 및 핵심 API 구현
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
...
  /me:
    get:
      tags: [Me]
      summary: 현재 사용자 정보 조회
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 사용자 정보
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: 인증 실패
  /me/sessions:
    get:
      tags: [Me]
      summary: 세션 목록 조회(옵션)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 성공
        '401':
          description: 인증 실패
components:
  schemas:
    UserStatus:
      type: string
      description: 사용자 상태
      enum: [ACTIVE, INACTIVE, LOCKED, DELETED]
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        status:
          $ref: '#/components/schemas/UserStatus'
        roles:
          type: array
          items:
            type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT